# public/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: GhostClass API
  version: 4.5.0
  description: |
    **GhostClass API** provides endpoints for authentication, profile synchronization,
    attendance integrations with EzyGo, telemetry, and build provenance.

    ## Authentication

    The API supports two authentication methods:

    1. **Bearer Token (CRON_SECRET)**: For automated cron jobs
       - Add `Authorization: Bearer <CRON_SECRET>` header
       - Used by automated sync processes

    2. **Session Cookie (SupabaseAuth)**: For authenticated users
       - Automatically included when logged in through the web interface
       - Managed by Supabase Auth

    ## Rate Limiting

    All endpoints are rate-limited to prevent abuse:

    - `/api/auth/save-token` & `/api/cron/sync`: Configurable via environment variables
    - `/api/backend/*`: Configurable via `PROXY_RATE_LIMIT_REQUESTS` and `PROXY_RATE_LIMIT_WINDOW`
    - Contact Form (Server Action): Configurable via environment variables

    Rate limit information is returned in response headers:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

    ## Base URLs

    - Production: Configured via `NEXT_PUBLIC_APP_URL`
    - Development: `http://localhost:3000`

    ## Error Handling

    All error responses follow a consistent format with appropriate HTTP status codes and descriptive messages.

  contact:
...

...
servers:
...
paths:
...
  /api/analytics/track:
    post:
      tags:
        - Analytics
      summary: Track a server-side analytics event
      security: []
      description: |
        Proxies GA4 Measurement Protocol events through the server so ad blockers
        do not suppress them. Validates event shapes, sanitises values to GA4 limits,
        and enforces origin and rate-limit guards before forwarding to GA4.

        **Authentication:** None required

        **Rate Limiting:** Configured via RATE_LIMIT_REQUESTS / RATE_LIMIT_WINDOW

      operationId: trackAnalyticsEvent

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - clientId
                - events
              properties:
                clientId:
                  type: string
                  maxLength: 100
                  description: GA4 client ID in `timestamp.randomhex` format
                  example: "1706108400.a1b2c3d4"
                events:
                  type: array
                  maxItems: 25
                  description: Up to 25 GA4 events per request
                  items:
                    type: object
                    required:
                      - name
                    properties:
                      name:
                        type: string
                        maxLength: 40
                        description: Event name (lowercase, underscores, digits only)
                      params:
                        type: object
                        description: Arbitrary GA4 event parameters
                        additionalProperties:
                          oneOf:
                            - type: string
                            - type: number
                            - type: boolean
                userProperties:
                  type: object
                  description: Optional GA4 user properties (key → string value)
                  additionalProperties:
                    type: string

      responses:
        "200":
          description: Events forwarded to GA4
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        "400":
          description: Validation error (bad clientId, malformed events, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden – request origin not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/auth/sync:
    post:
      tags:
        - Authentication
        - Mobile
      summary: Self-heal authentication state
      description: |
        Attempts to refresh Supabase session cookies and restore EzyGo auth token
        from server-side storage after client-side auth/session drift.
      operationId: syncAuth
      responses:
        "200":
          description: Sync attempt completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  terminal:
                    type: boolean
        "401":
          description: Session expired and cannot be recovered
          headers:
            x-terminal-auth:
              schema:
                type: string
              description: Present and set to `true` when re-authentication is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/register-fcm:
    post:
      tags:
        - Authentication
        - Mobile
      summary: Register mobile FCM push token
      description: |
        Registers or updates the Firebase Cloud Messaging (FCM) push token for the authenticated user.
        Enables sending push notifications to the user's mobile device.
      operationId: registerFcmToken
      security:
        - SupabaseAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fcm_token
              properties:
                fcm_token:
                  type: string
                  description: The Firebase Cloud Messaging registration token
                  example: "bk3RNwXXeDY:AP91bH..."
      responses:
        "200":
          description: FCM token registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        "400":
          description: Bad request - Missing IP or invalid JSON
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized - Invalid session or bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Validation failed - Empty token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message describing what went wrong
        message:
          type: string
          description: Detailed error message
      example:
        error: "Internal Server Error"
...
    RateLimitError:
      type: object
      properties:
        error:
          type: string
          description: Rate limit error message
          example: "Too many requests. Slow down!"
        retryAfter:
          type: integer
          description: Unix timestamp (in seconds) when the rate limit resets. Matches the `X-RateLimit-Reset` header.
          example: 1735689600
      required:
        - error
        - retryAfter
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer token authentication using CRON_SECRET environment variable.
        Used for automated cron jobs.

        Example: `Authorization: Bearer <CRON_SECRET>`

    SupabaseAuth:
      type: apiKey
      in: cookie
      name: sb-*
      description: |
        Session cookie authentication managed by Supabase Auth.
        Automatically included when users are logged in through the web interface.
        Cookie names typically start with 'sb-' prefix.
...